+++ src/soho/datagrid/soho-datagrid.component.ts
@@ -945,6 +945,25 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
   }
 
   /**
+   * The `emptyMessage` data grid option.
+   */
+  @Input() set emptyMessage(emptyMessage: SohoDataGridEmptyMessageOptions) {
+    /** Check for undefined/null and reset to the default message */
+    emptyMessage = emptyMessage ||
+      {title: (Locale ? Locale.translate('NoData') : 'No Data Available'), info: '', icon: 'icon-empty-no-data'};
+
+    this._gridOptions.emptyMessage = emptyMessage;
+    if (this.jQueryElement) {
+      this.datagrid.settings.emptyMessage = emptyMessage;
+      this.markForRefresh('emptyMessage', RefreshHintFlags.Rebuild);
+    }
+  }
+
+  get emptyMessage(): SohoDataGridEmptyMessageOptions {
+    return this._gridOptions.emptyMessage;
+  }
+
+  /**
    * Defines the source type of the grid, either:
    *
    * - "content-only" where table elements are provided in the body.
@@ -1004,6 +1023,9 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
 
   @Output()
   contextMenu = new EventEmitter<SohoDataGridRowClicked>();
+
+  @Output()
+  rowReordered = new EventEmitter<SohoDataGridRowReorderedEvent>();
 
   // -------------------------------------------
   // Host Bindings
@@ -1620,7 +1642,8 @@ export class SohoDataGridComponent implements OnInit, AfterViewInit, OnDestroy, 
       .on('rowdeactivated', (e: JQueryEventObject, args: SohoDataGridRowDeactivatedEvent) => { this.rowDeactivated.next(args); })
       .on('click', (e: JQueryEventObject, args: SohoDataGridRowClicked) => { this.rowClicked.next(args); })
       .on('dblclick', (e: JQueryEventObject, args: SohoDataGridRowClicked) => { this.rowDoubleClicked.next(args); })
-      .on('contextmenu', (e: JQueryEventObject, args: SohoDataGridRowClicked) => { this.contextMenu.next(args); });
+      .on('contextmenu', (e: JQueryEventObject, args: SohoDataGridRowClicked) => { this.contextMenu.next(args); })
+      .on('rowreorder', (e: JQueryEventObject, args: SohoDataGridRowReorderedEvent) => { this.rowReordered.next(args); });
   }
 
   /**