+++ src/soho/datagrid/soho-datagrid.d.ts
@@ -183,6 +183,7 @@ interface SohoDataGridOptions {
   userObject?: any;
 
   /**
+<<<<<<< HEAD
    * Optional callback called when a cell is rendered with the flag `postRender`
    * set to true.
    *
@@ -207,7 +208,73 @@ interface SohoDataGridOptions {
   * editors / formatters.
   */
   onEditCell?: SohoDataGridEditCellFunction;
+=======
+   * An empty message will be displayed when there are no rows in the grid.
+   * This accepts an object of the form SohoDataGridEmptyMessageOptions, set
+   * this to null for no message or it will default to 'No Data Found with an icon.'
+   */
+  emptyMessage?: SohoDataGridEmptyMessageOptions;
 }
+
+/**
+ * Example:
+ *
+ * { title: 'No Data Available',
+ *    info: 'Make a selection on the list above to see results',
+ *    icon: 'icon-empty-no-data',
+ *  button: {text: 'Run Query', cssClass: '' }
+ *   click: <function>}
+ * }
+ *
+ */
+interface SohoDataGridEmptyMessageOptions {
+  /**
+   * Title text.
+   */
+  title?: string;
+
+  /**
+   * Informational text.
+   */
+  info?: string;
+
+  /**
+   * Icon, to diplay (see svg-empty) - must be the full icon name.
+   */
+  icon?: string;
+
+  /**
+   * Button options - note this cannot be a Soho Angular Component.
+   */
+  button?: SohoDataGridEmptyMessageButtonOptions;
+>>>>>>> origin/4.3.3-rc
+}
+
+/**
+ * Button definition.
+ */
+interface SohoDataGridEmptyMessageButtonOptions {
+  /**
+   * Button text.
+   */
+  text: string;
+
+  /**
+   * Class of button.
+   */
+  cssClass?: string;
+
+  /**
+   * Element button for id.
+   */
+  id?: string;
+
+  /**
+   * Click function - @todo not working
+   */
+  click?: Function;
+}
+
 
 /**
  * Soho Data Grid Paging Options.
@@ -765,6 +832,13 @@ interface SohoDataGridRowRemoveEvent {
   target: any;
   value: any;
   oldValue: any;
+}
+
+interface SohoDataGridRowReorderedEvent {
+  start: any;
+  startIndex: number;
+  end: any;
+  endIndex: number;
 }
 
 interface SohoDataGridAddRowEvent {